+++ architecture/android/app/src/main/jni/DspFaust.h
@@ -29,7 +29,7 @@ public:
     void stop();
     bool isRunning();
 
-    unsigned long keyOn(int, int);
+    int keyOn(int, int);
     int keyOff(int);
     void propagateMidi(int, double, int, int, int, int);
 
@@ -40,8 +40,8 @@ public:
     void setParamValue(const char*, float);
     float getParamValue(const char*);
 
-    void setVoiceParamValue(const char*, unsigned long, float);
-    float getVoiceParamValue(const char*, unsigned long);
+    void setVoiceParamValue(const char*, int, float);
+    float getVoiceParamValue(const char*, int);
 
     const char* getParamAddress(int);
 
@@ -53,7 +53,6 @@ public:
 
     float getCPULoad();
     int getScreenColor();
-
 private:
 	FaustPolyEngine *fPolyEngine;
 };
